divide std coin test to another file
