+++ vm/src/tests/mod.rs
@@ -1,6 +1,7 @@
-pub mod simple_transaction_tests;
-pub mod vm_error_tests;
-pub mod table_tests;
-pub mod move_unit_tests;
 mod mock_chain;
 mod mock_tx;
+pub mod move_unit_tests;
+pub mod table_tests;
+pub mod tx_simple_tests;
+pub mod tx_std_coin_tests;
+pub mod vm_error_tests;
